Art Update
Added a ton of art assets, very few animations work atm, also added a
place holder background we can change it soon as we get a full art
background. Also added in 1 victim just to get a box collider working.
Also changed the color of victim display to white to contrast the
background currently. We might want a more legible font for, Think more
pixely cause the one we have now is not easily read